ts: Move CST namespace to parse-cst.d.ts, clarify node types
